Assignment0/requirements.txt
google-genai==1.36.0
trafilatura==2.0.0